Check list of all chuck pads to add from server.
This is the groundwork for multi-shred editing.  It checks the database
for all pads with a name like *-chuck-pad and gets their contents.

When they are addded, their local filename is changed to
running_*-chuck-pad.ck so we can know to only update, but not add to the
server.

This adds server initialization to the responsibilities of connect.sh,
and uses chuck --add to get files to a running --server --loop instance.